+++ packages/core/modal-dialog/examples/15-custom.js
@@ -1,6 +1,6 @@
 // @flow
 import React, { Component, type Node } from 'react';
-import styled from 'styled-components';
+import styled from '@emotion/styled';
 import Lorem from 'react-lorem-component';
 
 import Avatar from '@atlaskit/avatar';