+++ packages/core/modal-dialog/examples/65-long-headings.js
@@ -1,6 +1,6 @@
 // @flow
 import React, { Component } from 'react';
-import styled from 'styled-components';
+import styled from '@emotion/styled';
 import Lorem from 'react-lorem-component';
 
 import Button, { ButtonGroup } from '@atlaskit/button';