+++ packages/editor/editor-mobile-bridge/src/__tests__/integration/native-to-web/links.ts
@@ -24,7 +24,7 @@ BrowserTestCase(
       browser,
       'onLinkUpdate',
       'Atlassian',
-      'https://www.atlassian.com',
+      'https://www.google.com',
     );
 
     const doc = await browser.$eval(editable, getDocFromElement);
@@ -49,7 +49,7 @@ BrowserTestCase(
       browser,
       'onLinkUpdate',
       'link',
-      'https://www.atlassian.com',
+      'https://www.google.com',
     );
 
     const doc = await browser.$eval(editable, getDocFromElement);
@@ -77,7 +77,7 @@ BrowserTestCase(
       browser,
       'onLinkUpdate',
       'link',
-      'https://www.atlassian.com',
+      'https://www.google.com',
     );
 
     const doc = await browser.$eval(editable, getDocFromElement);
@@ -99,7 +99,7 @@ BrowserTestCase(
       browser,
       'onLinkUpdate',
       'Atlassian',
-      'https://www.atlassian.com',
+      'https://www.google.com',
     );
 
     await browser.type(editable, [...times(4, constant('ArrowLeft'))]);
@@ -108,7 +108,7 @@ BrowserTestCase(
       browser,
       'onLinkUpdate',
       'This is Atlassian',
-      'https://www.atlassian.com',
+      'https://www.google.com',
     );
 
     const doc = await browser.$eval(editable, getDocFromElement);
@@ -130,7 +130,7 @@ BrowserTestCase(
       browser,
       'onLinkUpdate',
       'Google',
-      'https://www.atlassian.com',
+      'https://www.google.com',
     );
 
     await browser.type(editable, [...times(4, constant('ArrowLeft'))]);
@@ -167,7 +167,7 @@ BrowserTestCase(
       browser,
       'onLinkUpdate',
       'link',
-      'https://www.atlassian.com',
+      'https://www.google.com',
     );
 
     await browser.type(editable, [