+++ packages/jira/refinement-bar/src/fields/IssueSelect/Controller.js
@@ -1,0 +1,7 @@
+// @flow
+
+import SelectController from '../Select/Controller';
+
+const IssueSelectController = SelectController;
+
+export default IssueSelectController;