+++ packages/jira/refinement-bar/src/fields/index.js
@@ -1,0 +1,11 @@
+// @flow
+
+export { default as AsyncSelectFilter } from './AsyncSelect';
+export { default as AvatarSelectFilter } from './AvatarSelect';
+export { default as AvatarAsyncSelectFilter } from './AvatarAsyncSelect';
+export { default as IssueSelectFilter } from './IssueSelect';
+export { default as IssueAsyncSelectFilter } from './IssueAsyncSelect';
+export { default as NumberFilter } from './Number';
+export { default as SelectFilter } from './Select';
+export { default as SearchFilter } from './Search';
+export { default as TextFilter } from './Text';