+++ packages/media/media-card/src/index.ts
@@ -15,13 +15,6 @@ import { MediaViewerDataSource } from '@atlaskit/media-viewer';
 // the only components we expose to consumers is Card and CardView
 export { default as Card } from './root/card/cardLoader';
 
-export { CardView } from './root/cardViewLoader';
-
-export {
-  CardViewState,
-  CardViewOwnProps as CardViewProps,
-} from './root/cardView';
-
 export * from './actions';
 
 export type CardStatus =
@@ -146,4 +139,4 @@ export interface CardState {
   error?: Error;
 }
 
-export { defaultImageCardDimensions } from './utils';
+export { defaultImageCardDimensions, CardLoading, CardError } from './utils';