+++ packages/media/media-test-helpers/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atlaskit/media-test-helpers
 
+## 21.2.2
+- Updated dependencies [0ff405bd0f](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/0ff405bd0f):
+  - @atlaskit/media-core@29.1.2
+  - @atlaskit/media-store@11.0.5
+  - @atlaskit/media-card@60.0.0
+  - @atlaskit/media-picker@38.1.3
+
 ## 21.2.1
 - [patch] [d13fad66df](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d13fad66df):